Фикс параметрво полей в  таблице  цен. Изменено с smallint 5 на int 10

+++ var/dump/m2_item_price.strc.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `m2_item_price` (
-  `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
-  `item_id` smallint(5) unsigned NOT NULL,
-  `order` smallint(5) unsigned NOT NULL,
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
+  `item_id` int(11) unsigned NOT NULL,
+  `order` int(11) unsigned NOT NULL,
   `type` varchar(32) NOT NULL,
   `content` mediumtext NOT NULL,
   `price_value` varchar(255) NOT NULL,
